+++ redux/types.js
@@ -6,6 +6,7 @@ export const EMPLOYEE_LOGOUT = 'EMPLOYEE_LOGOUT';
 
 // // Employee Activities
 
+export const GET_INTERNET_CONNECTION_STATUS = 'GET_INTERNET_CONNECTION_STATUS';
 export const GET_WORKSPACES_LIST = 'GET_WORKSPACES_LIST';
 export const GET_PROJECTS_LIST = 'GET_PROJECTS_LIST';
 export const GET_TASKS_LIST = 'GET_TASKS_LIST';
